Simplify luma downsampling for wienerns cross_filt

See merge request AOMediaCodec/avm!469

+++ av1/common/thread_common.c
@@ -858,7 +858,7 @@ static void foreach_rest_unit_in_planes_mt(AV1LrStruct *lr_ctxt,
 
 #if CONFIG_WIENER_NONSEP && CONFIG_WIENER_NONSEP_CROSS_FILT
   uint8_t *luma = NULL;
-  uint8_t *luma_buf;
+  uint16_t *luma_buf;
   const YV12_BUFFER_CONFIG *dgd = &cm->cur_frame->buf;
   int luma_stride = dgd->crop_widths[1] + 2 * WIENERNS_UV_BRD;
   luma_buf = wienerns_copy_luma_highbd(
